LocalLLMWithOllama/requirements.txt
fastapi
uvicorn
requests
python-dotenv
gradio